==> 1_filter_data.ipynb <==
!find data/2018BlackLives/ -name '*.json' -exec cat {} \; > data/2018_black_lives_matter.json

!find data/2018GunControl/ -name '*.json' -exec cat {} \; > data/2018_gun_control.json

!find data/2018ClimateChange/ -name '*.json' -exec cat {} \; > data/2018_climate_change.json

!find data/2018MeToo/ -name '*.json' -exec cat {} \; > data/2018_me_too.json

file= "data/2018_me_too.json"

!sed -i 's/$/,/' {file}

!sed -i '$ s/.$//' {file}

!printf "]" >> {file}

!sed -i '1s/^/[ /' {file}

!sed -i 's/\ud83c//g'  {file}

!sed -i 's/\ud83d//g'  {file}

!sed -i 's/\ud83e//g'  {file}

!sed -i 's/\ud835//g'  {file}

!sed -i 's/\ud80c//g'  {file}

!sed -i 's/\udb40//g'  {file}

!sed -i 's/\ud84f//g'  {file}

!sed -i 's/\uda09//g'  {file}

!sed -i 's/\x00//g'  {file}

!sed -i 's/\ud803//g'  {file}

!sed -i 's/\udbb9//g'  {file}

!sed -i 's/\ud81c//g'  {file}

!sed -i 's/\ud800//g'  {file}

!sed -i 's/\uda32//g'  {file}

file
