==> 4_Frequency.ipynb <==
%reload_ext cypher
import matplotlib
import pandas as pd
import matplotlib.pyplot as plt
import datetime
%matplotlib inline
matplotlib.rcParams['figure.figsize'] = (12.0, 4.0)

## Number of Tweets by day

tweets = pd.read_csv("csvs/tweets.csv", sep="\t")
times = tweets.created_at.apply(lambda x: datetime.datetime.strptime(x.split(" ", 1)[1], "%b %d %H:%M:%S +0000 %Y") if type(x) == str else None)

dt_ndx = pd.DatetimeIndex(times)
sr = pd.Series([1] * len(dt_ndx), index=dt_ndx)
daily_count = sr.resample('1D').apply('sum')
daily_count = daily_count.iloc[-31:]

daily_count.sort_index(ascending = False, inplace=True)
frame = daily_count.to_frame()
frame.columns = ['Number of tweets']
frame.index.names = ['Date']
frame

## Number of tweets by days

daily_count.sort_values(ascending = False, inplace=True)
frame = daily_count.to_frame()
frame.columns = ['Number of tweets']
frame.index.names = ['Date']
frame

## Top 20 hours with more tweets. 

dt_ndx = pd.DatetimeIndex(times)
sr1 = pd.Series([1] * len(dt_ndx), index=dt_ndx)
hourly_count = sr1.resample('1H').apply('sum')
hourly_count = hourly_count.iloc[-31:]

hourly_count.sort_values(ascending = False, inplace=True)
hdf = hourly_count.to_frame()
hdf.columns = ['Number of tweets']
hdf.index.names = ['Hours']
hdf.head(20)
