==> 9_Topics.ipynb <==
## Topics 

from __future__ import print_function
from time import time

from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.decomposition import NMF
from nltk.corpus import stopwords
import pandas as pd
%reload_ext cypher
import re

stop_words_ = set(stopwords.words('english'))  
new_set = set(['000', 'de', 'rt', 'http', 'https', 'amp', '1', '25', 'pm', '2', 'rtudf0a', 'udf0a', 'lifeudc68udfff', 'udfa4udc69udffb', 'udfa8', 'udf08udf37udf3audd16', 'lmaooooooooooooooooooooooooooooooooooooooooooooude02ude02', 'udf31small', '' ])
stop_words_ = stop_words_.union(new_set)

## Topics by NMF

n_top_words = 10
n_features = 10000
n_components = 20
full_topics= {}

def train_NMF(data_samples):
    tfidf_vectorizer = TfidfVectorizer(max_df=1.0, min_df=1,
                                   max_features=n_features,
                                   stop_words=stop_words_)
    tfidf = tfidf_vectorizer.fit_transform(data_samples)

    nmf = NMF(n_components=n_components, random_state=1,
          alpha=.1, l1_ratio=.5).fit(tfidf)
    
    feature_names = tfidf_vectorizer.get_feature_names()
   
    print_top_words(nmf.components_, feature_names, n_top_words)
    
    return nmf, tfidf_vectorizer
      
def print_top_words(model, feature_names, n_top_words):
    for topic_idx, topic in enumerate(model):
        full_topics[topic_idx] = " ".join([feature_names[i] for i in topic.argsort()[:-n_top_words - 1:-1]])
        print("Topic #%d: " % topic_idx + full_topics[topic_idx] )
        print()
        
def clean_tweets(tweets):
    patternUrl = re.compile('http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+')
    patternUsers =  re.compile('@(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+')
    patternTags = re.compile('#(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+')
    allTweets = pd.Series(tweets).str.cat(sep='\n|\n||\n')
    allTweets = patternUrl.sub('', allTweets) #removing urls 
    allTweets = patternUsers.sub('', allTweets) #removing users
    allTweets = patternTags.sub('', allTweets) #removing tags
    return allTweets.split('\n|\n||\n')

ecos = [ 'BLM', 'MT', 'CCH', 'GC']
for eco in ecos:
    tweetsQ = %cypher match  (n:tweet)<-[r :TWEETS]-(n2:user) where n.eco = '{eco}' and n.lang='en' return substring(n.text, 0, 10000000) as text
    training_data = clean_tweets(tweetsQ.get_dataframe().text.unique())
    print(eco, len(training_data), 'tweets -------------------------------------------------------------------')
    nmf, tfidf_vectorizer = train_NMF(training_data)
