==> 10_Objetificacion_Lenguaje_main.ipynb <==
import pandas as pd

## Threads - Conversaciones

#empezar en 13001

df = pd.read_csv("conversations/h5.csv", sep="\t", header=None)
df2 = pd.read_csv("conversations/h6.csv", sep="\t", header=None)
df3 = pd.read_csv("conversations/h7.csv", sep="\t", header=None)
df = pd.concat([df, df2, df3])
df.columns = ['Twitter ID', 'Path', 'Path_lenght', 'Relationships', 'Date Start', 'Date End', 'Nodes IDs']
df.head()

## Stats on the length of the Conversations. 

len(df)

df_conversations  = pd.DataFrame(columns=['N', 'Mean', 'Standard Dev', 'Min', 'Max'])

df_conversations.loc['Conversations Paths'] = [len(df), df['Path_lenght'].mean(), df.loc[:,"Path_lenght"].std(), min(df.loc[:,"Path_lenght"]), max(df.loc[:,"Path_lenght"])]

df_conversations

## Stats on Retweets, Comments (Quotes + Replies), Likes

import re
def get_observations(labels, df):
    observations = [] 
    for index, row in df.iterrows():
        n=0
        if 'LIKES' in labels:
            print (row['Path'])
            n = get_likes(row['Path'])
            break
        else: 
            for label in labels:
                if pd.notnull(row['Relationships']):
                     n = n + (row['Relationships'].count(label))
        observations.append(n)
    return observations

def get_likes(path):
    n = 0
    pos = [m.start() for m in re.finditer(", 'likes': ", a)]
    for i in pos:
        start = a.find('\'',i+11) +1 
        end = a.find('\'',start)
        n = int(a[start:end])
        print(n)

a = "[{'created_at': 'Wed May 30 14:02:28 +0000 2018', 'eco': 'MT', 'likes': '0', 'lang': 'fa', 'tid': '1001826179899449344', 'text': 'RT @G0NEWITHTHEWlND: شیشه میکشید.سال 91. همین موقع ها.. امتحان جبرو احتمال داشتم. ساعت 3صبح خوابیدم برای 8که امتحان داشتم چشمام که سنگین شد…'}, {'eco': 'MT'}, {'created_at': 'Wed May 30 04:21:30 +0000 2018', 'eco': 'MT', 'likes': '4273', 'lang': 'fa', 'tid': '1001679970761166848', 'text': 'شیشه میکشید.سال 91. همین موقع ها.. امتحان جبرو احتمال داشتم. ساعت 3صبح خوابیدم برای 8که امتحان داشتم چشمام که سنگین شد یهو اومد بالای سرم و شروع کرد به بوسیدن سینه هام.. فقط تونستم جیغ بکشم .. هیچی نگفت. فحش داد رفت بیرون.. پدرم !پدر واقعی خودم ! ناپدری نه !!  پدر!!  #MeToo'}, {'created_at': 'Wed May 30 14:02:28 +0000 2018', 'eco': 'MT', 'likes': '0', 'lang': 'fa', 'tid': '1001826179899449344', 'text': 'RT @G0NEWITHTHEWlND: شیشه میکشید.سال 91. همین موقع ها.. امتحان جبرو احتمال داشتم. ساعت 3صبح خوابیدم برای 8که امتحان داشتم چشمام که سنگین شد…'}, {'eco': 'MT'}, {'created_at': 'Wed May 30 04:21:30 +0000 2018', 'eco': 'MT', 'likes': '4273', 'lang': 'fa', 'tid': '1001679970761166848', 'text': 'شیشه میکشید.سال 91. همین موقع ها.. امتحان جبرو احتمال داشتم. ساعت 3صبح خوابیدم برای 8که امتحان داشتم چشمام که سنگین شد یهو اومد بالای سرم و شروع کرد به بوسیدن سینه هام.. فقط تونستم جیغ بکشم .. هیچی نگفت. فحش داد رفت بیرون.. پدرم !پدر واقعی خودم ! ناپدری نه !!  پدر!!  #MeToo'}]"

import re
pos = [m.start() for m in re.finditer(", 'likes': ", a)]
for i in pos:
    start = a.find('\'',i+11) +1 
    end = a.find('\'',start)
    n = int(a[start:end])
    print(n)

#df = df.append(pd.DataFrame(get_observations(['REPLIES_TO', 'QUOTED_TO'], df), columns=['Comments']), sort=False)
df = df.append(pd.DataFrame(get_observations(['LIKES'], df), columns=['Likes']), sort=False)
df = df.append(pd.DataFrame(get_observations(['RETWEETS'], df), columns=['Retweets']), sort=False)

df.head()

def get_stats(labels, df):
    #['N', 'Mean', 'Standard Dev', 'Min', 'Max']
    stats = [] 
    for label in labels:
        if label == 'LIKES':
            return getLikes()
        else:

            for a in df['Relationships']:
               
        return stats

df_relations  = pd.DataFrame(columns=['N', 'Mean', 'Standard Dev', 'Min', 'Max'])
df_relations.loc['Comments'] = get_stats(['REPLIES_TO', 'QUOTED_TO'], df)
df_relations.loc['Likes'] = get_stats(['LIKES'], df)
df_relations.loc['Retweets'] = get_stats(['RETWEETS'], df)
df_relations
