==> 10_Objetificacion_Lenguaje_Parsing_Ids.ipynb <==
%reload_ext cypher
import pandas as pd

## Threads - Conversaciones

tweets = %cypher MATCH (n:tweet)-[r :RETWEETS|REPLIES_TO|QUOTED_TO*]-> (m:tweet) return distinct (n.tid) as tid

tweets_list = list(tweets.get_dataframe().tid)
len(tweets_list)

#all_paths_df =  pd.DataFrame(columns=['tweet_id', 'path', 'path_len', 'relationships', 'hours_diff', 'ids']) 

len(tweets_list)

with open('conversations/tweets_ids.txt', 'w') as f:
    for item in tweets_list:
        f.write("%s\n" % item)

with open('conversations/tweets_ids.txt') as f:
    tweets_list = f.read().splitlines()

len(tweets_list)

%cypher MATCH path=(n:tweet)-[r :RETWEETS|REPLIES_TO|QUOTED_TO*]-> (m:tweet) WHERE n.tid = '1002036506050088960' RETURN n.tid as tweet_id, path, length(path) as path_len, [r in relationships(path) | type(r)] as relationships, [t in nodes(path)| t.tid] as nodes ORDER BY path_len DESC LIMIT 1
